moved CommitementProof deser/ser
